[#6 - Advanced Multithreading] fix mentor's remarks

+++ m6-advanced-multithreading/t6-recursivetask-recursiveaction/src/test/java/pl/mentoring/recursive/FibonacciTaskTest.java
@@ -10,14 +10,14 @@ class FibonacciTaskTest {
 
     @Test
     void compute() {
-        assertEquals(1134903170L, new ForkJoinPool().invoke(new FibonacciTask(45)).longValue());
+        assertEquals(1134903170L, new ForkJoinPool().invoke(new FibonacciTask(45)));
     }
 
     @Test
     void compute_shouldReturnExpectedValue() {
         ForkJoinPool pool = new ForkJoinPool();
-        assertEquals(1L, pool.invoke(new FibonacciTask(2)).longValue());
-        assertEquals(8L, pool.invoke(new FibonacciTask(6)).longValue());
-        assertEquals(6765L, pool.invoke(new FibonacciTask(20)).longValue());
+        assertEquals(1L, pool.invoke(new FibonacciTask(2)));
+        assertEquals(8L, pool.invoke(new FibonacciTask(6)));
+        assertEquals(6765L, pool.invoke(new FibonacciTask(20)));
     }
 }
